Replace the inline script for alert_div and alert_messages with a new almost-pure-JavaScript approach Split flash helper methods out to their own helper module Simplify render_as_list Move bool_icon and bool_text over to the formatting helper
