+++ app/javascript/sweetalert/index.js
@@ -1,5 +1,7 @@
 import swal from 'sweetalert2/dist/sweetalert2.js'
 import Rails from "@rails/ujs";
+import './alerts'
+
 window.Swal = swal;
 
 const Toast = Swal.mixin({
@@ -23,9 +25,11 @@ const PersistentToast = Swal.mixin({
   timerProgressBar: false
 })
 
+// Make these methods available to the window object, so that they can be used in inline scripts.
 window.Toast = Toast;
 window.PersistentToast = PersistentToast;
 
+// TODO: Doubt this works this way anymore? Check.
 Rails.confirm = function (message, element) {
   const swalWithBootstrap = swal.mixin({
     buttonsStyling: true,
@@ -48,4 +52,3 @@ Rails.confirm = function (message, element) {
       }
     });
 }
-
