--- app/views/layouts/_alert_message.erb
@@ -1,18 +1,0 @@
-<% alert_type = swal_alert_info(key) %>
-
-<% if messages.count <= 1 
-    message = messages.first 
-  else
-    message = render_as_list(messages, 'ul')
-  end %>
-
-<% # Awaits make the code wait until the alert has disappeared before continuing. If you do not do this, only one alert will be visible. %>
-<% # The surrounding <script> tag and async function is in the alert_div partial. If you use this elsewhere, make sure to wrap calls to the partial in a script tag and async function. %>
-<% case alert_type %>
-  <% when 'success' %>
-      await showToast('<%= alert_type %>', '<%= " #{message}".html_safe %>');
-  <% when 'error' %>
-      await showError('<%= " <div class=\"text-justify\">#{message}</div>".html_safe %>');
-  <% else # Meant for info, but also used as a catch in case there is no flash category.%>
-      await showPersistentToast('<%= alert_type %>', '<%= " #{message}".html_safe %>');
-<% end %>
